% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centering.R
\name{D_center}
\alias{D_center}
\title{Double centering}
\usage{
D_center(A)
}
\arguments{
\item{A}{dist object or data matrix}
}
\value{
\code{D_center} returns double centered distance matrix
}
\description{
Double centered version of distance matrix of A
}
\examples{
A <- iris[1:50, 1:4]
A_Dcenter <- D_center(A)

}
\references{
Park, Trevor, Xiaofeng Shao, and Shun Yao. "Partial martingale difference correlation." Electronic Journal of Statistics 9.1 (2015): 1492-1517.
}
