\name{MDD-package}
\alias{MDD-package}
\alias{MDD}
\docType{package}
\title{
\packageTitle{MDD}
}
\description{
\packageDescription{MDD}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MDD}
\packageIndices{MDD}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MDD}

Maintainer: \packageMaintainer{MDD}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
