% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd_stats.R
\name{mdd}
\alias{mdd}
\title{Martingale Difference Divergence and Correlation Statistics}
\usage{
mdd(X, Y)
}
\arguments{
\item{X}{distances or data of first sample}

\item{Y}{distances or data of second sample}
}
\value{
\code{mdd} returns a list containing
\itemize{
  \item{MDD: }{sample martingale difference divergence}
  \item{MDC: }{sample martingale difference correlation}
}
}
\description{
Sample martingale difference divergence and sample martingale difference correlation between X and Y
}
\examples{
A <- iris[1:50, 1:2]
B <- iris[1:50, 3:4]
mdd(A, B)

}
\references{
Shao, Xiaofeng, and Jingsi Zhang. "Martingale difference correlation and its use in high-dimensional variable screening." Journal of the American Statistical Association 109.507 (2014): 1302-1318.
}
