% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd_test_bootstrap.R
\name{mdd_test_boot}
\alias{mdd_test_boot}
\title{Conditional mean independence test using wild bootstrap}
\usage{
mdd_test_boot(X, Y, B = 999)
}
\arguments{
\item{X}{distances or data of first sample}

\item{Y}{distances or data of second sample}

\item{B}{number of replicates}
}
\value{
\code{mdd_test_boot} returns a list containing
\itemize{
  \item{test_statistic: }{observed value of the test statistic}
  \item{MDD_sq_U: }{unbiased estimate of squared martingale difference divergence}
  \item{replicates: }{replicates of the test statistic}
  \item{p_value: }{approximate p-value of the test}
}
}
\description{
Conditional mean independence test using wild bootstrap
}
\examples{
A <- iris[1:50, 1:2]
B <- iris[1:50, 3:4]
set.seed(1)
mdd_test_boot(A, B)

}
\references{
Chung Eun Lee, Xianyang Zhang, and Xiaofeng Shao. Testing the conditional mean independence for functional data. Biometrika, forthcoming, 2019.
}
