% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd_test_permutation.R
\name{mdd_test_perm}
\alias{mdd_test_perm}
\title{Conditional mean independence test using permutation}
\usage{
mdd_test_perm(X, Y, R = 999)
}
\arguments{
\item{X}{distances or data of first sample}

\item{Y}{distances or data of second sample}

\item{R}{number of replicates}
}
\value{
\code{mdd_test_perm} returns a list containing
\itemize{
  \item{test_statistic: }{observed value of the test statistic}
  \item{MDD: }{sample martingale difference divergence}
  \item{replicates: }{replicates of the test statistic}
  \item{p_value: }{approximate p-value of the test}
}
}
\description{
Conditional mean independence test using permutation
}
\examples{
A <- iris[1:50, 1:2]
B <- iris[1:50, 3:4]
set.seed(1)
mdd_test_perm(A, B)

}
\references{
Shao, Xiaofeng, and Jingsi Zhang. "Martingale difference correlation and its use in high-dimensional variable screening." Journal of the American Statistical Association 109.507 (2014): 1302-1318.

Park, Trevor, Xiaofeng Shao, and Shun Yao. "Partial martingale difference correlation." Electronic Journal of Statistics 9.1 (2015): 1492-1517.
}
