% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd_U_stats.R
\name{mdd_U}
\alias{mdd_U}
\title{Unbiased martingale difference divergence statistics}
\usage{
mdd_U(X, Y)
}
\arguments{
\item{X}{distances or data of first sample}

\item{Y}{distances or data of second sample}
}
\value{
\code{mdd_U} returns a list containing
\itemize{
  \item{MDD_sq_U: }{unbiased estimate of squared martingale difference divergence}
  \item{MDC_sq_U: }{bias-corrected estimate of squared martingale difference correlation}
}
}
\description{
Unbiased estimate of squared martingale difference divergence and bias-corrected estimate of squared martingale difference correlation
}
\examples{
A <- iris[1:50, 1:2]
B <- iris[1:50, 3:4]
mdd_U(A, B)

}
\references{
Park, Trevor, Xiaofeng Shao, and Shun Yao. "Partial martingale difference correlation." Electronic Journal of Statistics 9.1 (2015): 1492-1517.
}
