% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centering.R
\name{U_center}
\alias{U_center}
\title{U-centering}
\usage{
U_center(A)
}
\arguments{
\item{A}{dist object or data matrix}
}
\value{
\code{U_center} returns U-centered distance matrix
}
\description{
U-centered version of distance matrix of A
}
\examples{
A <- iris[1:50, 1:4]
A_Ucenter <- U_center(A)

}
\references{
Park, Trevor, Xiaofeng Shao, and Shun Yao. "Partial martingale difference correlation." Electronic Journal of Statistics 9.1 (2015): 1492-1517.
}
