% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_innerproduct.R
\name{U_innerproduct}
\alias{U_innerproduct}
\title{Inner product in the Hilbert space of U-centered distance matrices}
\usage{
U_innerproduct(A, B = NULL)
}
\arguments{
\item{A}{U-centered distance matrix}

\item{B}{U-centered distance matrix}
}
\value{
\code{U_innerproduct} returns the inner product of A and B, which is a scalar
}
\description{
Inner product in the Hilbert space of U-centered distance matrices
}
\examples{
A <- iris[1:50, 1:2]
B <- iris[1:50, 3:4]
A_Ucenter <- U_center(A)
B_Ucenter <- U_center(B)
U_innerproduct(A_Ucenter, B_Ucenter)

}
\references{
Park, Trevor, Xiaofeng Shao, and Shun Yao. "Partial martingale difference correlation." Electronic Journal of Statistics 9.1 (2015): 1492-1517.
}
